+++ CHANGES.md
@@ -4,7 +4,7 @@
 Changes in 4.5
 --------------
 - Add Calendar header hook in Calendar.php
-- Add Calendar Day functions in CalendarDay.inc.php
+- Add Calendar Day functions in CalendarDay.inc.php & Calendar.php
 
 Changes in 4.4
 --------------